Fix: Correct printf format specifier in PONG logging
- Change %zu to %u with cast to (unsigned int)
- Fixes "size=zu" display in logs
- Now shows actual packet size correctly
- Cosmetic fix, does not affect functionality

+++ SwordOnline/Sources/MultiServer/GameServer/KSOServer.cpp
@@ -2161,8 +2161,8 @@ void KSwordOnLineSever::PlayerMessageProcess(const unsigned long lnID, const cha
 		{
 			if (*(BYTE*)pData == c2s_ping)
 			{
-				printf("[GS-RECV-PONG] lnID=%lu received PONG packet (size=%zu), calling ProcessPingReply...\n",
-					   lnID, dataLength);
+				printf("[GS-RECV-PONG] lnID=%lu received PONG packet (size=%u), calling ProcessPingReply...\n",
+					   lnID, (unsigned int)dataLength);
 				ProcessPingReply(lnID, pData, dataLength);
 			}
 			else if (*(BYTE*)pData == c2s_extendtong)
